% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_reads.R
\name{bam_reads}
\alias{bam_reads}
\title{Returns the number of distinct reads in the BAM file alignment. This may be different from
`bam_length`, in particular if there are chimeric alignments, as those will present as multiple records
despite being a single read alignment. Read pairs are counted as 2 distinct reads.}
\usage{
bam_reads(bamfile, checks = 3)
}
\arguments{
\item{bamfile}{path to the BAM file}

\item{checks}{1 indicates checking only for SA tag, 2 only for rname, and 3 for both. Default is both.}
}
\description{
How chimeric reads are defined is described in `chimeric_reads`.
}
\details{
Planned improvements: Because Rsamtools does not have the capability for the 0x800 flag, implement
as a function based on the RSeqAn library instead for easier flagging of chimeric reads.
}
