% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_reads.R
\name{bam_records}
\alias{bam_records}
\title{Returns the number of alignment records in the BAM file alignment.}
\usage{
bam_records(bamfile)
}
\arguments{
\item{bamfile}{path to the BAM file}
}
\description{
Returns the number of alignment records in the BAM file alignment.
}
