% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimeric_reads.R
\name{chimeric_proportion}
\alias{chimeric_proportion}
\title{Returns proportion of reads that have been mapped to two distinct portions of the genome, here
defined as different chromosomes or sequence segments.}
\usage{
chimeric_proportion(bamfile, checks = 3)
}
\arguments{
\item{bamfile}{the path to the BAM file}

\item{checks}{1 indicates checking only for SA tag, 2 only for rname, and 3 for both. Default is both.}
}
\description{
How chimeric reads are identified is described in `chimeric_reads`. The indices returned will be
the proportion of reads with any of the features above. The user also has
the option to only check for 1 of the 2.
}
\details{
NOTE that the absence of either of the 2 features above does NOT mean that your alignment does not have
chimeras, especially if you are using older versions of reference aligners. This is merely a
heuristic diagnostic using convenient metrics to assess overall quality of your alignment.

Planned improvements:
Also check for if the 0x800 flag exists, indicating a supplementary and thus chimeric alignment.
This flag only exists in version 1.5 of the SAM and BAM formats and beyond.
}
