% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimeric_reads.R
\name{chimeric_rname}
\alias{chimeric_rname}
\title{Identifies reads with the same qname but different rname from a bamfile.}
\usage{
chimeric_rname(bamfile)
}
\arguments{
\item{bamfile}{the path to the BAM file}
}
\value{
list of index record pairs that have the same qname but different rname
}
\description{
Identifies reads with the same qname but different rname from a bamfile.
}
