% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimeric_reads.R
\name{chimeric_tags}
\alias{chimeric_tags}
\title{Identifies reads with the SA tag, indicating a chimeric read as defined by bwa-mem.}
\usage{
chimeric_tags(bamfile)
}
\arguments{
\item{bamfile}{the path to the BAM file}
}
\description{
Identifies reads with the SA tag, indicating a chimeric read as defined by bwa-mem.
}
