% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimeric_reads.R
\name{num_chimeric}
\alias{num_chimeric}
\title{Return number of chimeric reads from a bamfile.}
\usage{
num_chimeric(bamfile, checks = 3)
}
\arguments{
\item{bamfile}{the path to the BAM file}

\item{checks}{1, 2, or 3.}
}
\description{
Return number of chimeric reads from a bamfile.
}
