% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_bam_or_sam}
\alias{read_bam_or_sam}
\title{Read in a BAM or SAM file. For a BAM file package will need to be
linked against zlib.}
\usage{
read_bam_or_sam(bamFileName)
}
\arguments{
\item{fileName}{path to SAM or BAM file}
}
\description{
Read in a BAM or SAM file. For a BAM file package will need to be
linked against zlib.
}
